fix: createdAt/Cid-type 수정
createdAt의 생성 방식을 통합하였고, cid의 타입을 Long으로 모두 통합하였습니다.

+++ src/main/java/com/github/backendpart/repository/ProductRepository.java
@@ -3,5 +3,5 @@ package com.github.backendpart.repository;
 import com.github.backendpart.web.entity.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
 }
